fix: drop IGNORE scoring; omit defaults in synth harness
Do not soft-score ungrounded optionals as verify=ignore. Keep gold clean
via labeler prompt rules + deterministic OMIT_FROM_GOLD strip in parse/
format harness. Scoring still uses llmAsAJudge for internet lookup and
nonempty for conversation titles.
