+++ ts/packages/benchmarks/src/translationBench/synthesizer/index.ts
@@ -15,3 +15,4 @@ export * from "./synthesizerPrompts.js";
 export * from "./utteranceDisambiguation.js";
 export * from "./catalogGenerator/index.js";
 export { seedQaJsonlAdapter } from "./adapters/seedQaJsonlAdapter.js";
+export * from "./goldParameterHygiene.js";
