+++ ts/packages/benchmarks/src/translationBench/synthesizer/quality-verifier.prompt.yaml
@@ -29,7 +29,7 @@ format_checker:
     - history_shape_when_present
     - active_schema_membership
     - utterance_action_disambiguation
-    - strip_ungrounded_optional_gold_parameters
+    - strip_omit_from_gold_parameters
 
 # Stage 2 — semantic / quality judge (LLM)
 semantic_checker:
